requirements.txt
geopandas
shapely
numpy
requests
python-dotenv
